backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
bcrypt==4.0.1
pyotp==2.9.0
qrcode[pil]==7.4.2
PyJWT==2.8.0
Pillow==9.5.0
gunicorn==21.2.0 